requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
matplotlib
tensorflow
wordcloud
joblib
comet-ml
dvc